About Us & other fixes
I fixed the carousel, so now it works everytime. I also fixed the size
so it’s always 100% width and 800px.

I also moved the placeholder image to a folder called “img” in our
“Test2”.

I also finished the “About Us” site and put all our pretty faces on it.

I also put the “footerfoot” on all of the pages and removed the
<footer> in login.html